chore: trigger new release to pull in latest snakemake
Currently, the snakemake-github-action seems to pull in a broken version of snakemake, with a bug that has since been fixed. So simply pulling in the latest release should fix this.

Current error message when using module import:
```
Rule convert_pfam is not defined in this workflow.
```

See e.g. here:
https://github.com/snakemake-workflows/rna-seq-kallisto-sleuth/actions/runs/5865262889/job/15901825940#step:4:18